Restructuring files to fit more common standards and cleaning password relationships.
