+++ .idea/workspace.xml
@@ -5,11 +5,18 @@
   </component>
   <component name="ChangeListManager">
     <list default="true" id="257b68f9-7575-40dc-9d1c-f6cbe7a3603c" name="Changes" comment="">
+      <change beforePath="$PROJECT_DIR$/.idea/hindsite.iml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/hindsite.iml" afterDir="false" />
       <change beforePath="$PROJECT_DIR$/.idea/workspace.xml" beforeDir="false" afterPath="$PROJECT_DIR$/.idea/workspace.xml" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/main.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/main.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/tables/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/tables/__init__.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/tables/password.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/tables/password.py" afterDir="false" />
-      <change beforePath="$PROJECT_DIR$/src/tables/user.py" beforeDir="false" afterPath="$PROJECT_DIR$/src/tables/user.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/app/static/bootstrap.min.css" beforeDir="false" afterPath="$PROJECT_DIR$/app/static/bootstrap.min.css" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/app/static/htmx.min.js" beforeDir="false" afterPath="$PROJECT_DIR$/app/static/htmx.min.js" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/app/templates/base.html" beforeDir="false" afterPath="$PROJECT_DIR$/app/templates/base.html" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/app/templates/index.html" beforeDir="false" afterPath="$PROJECT_DIR$/app/templates/index.html" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/main.py" beforeDir="false" afterPath="$PROJECT_DIR$/app/main.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/tables/__init__.py" beforeDir="false" afterPath="$PROJECT_DIR$/tables/__init__.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/tables/group.py" beforeDir="false" afterPath="$PROJECT_DIR$/tables/group.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/tables/password.py" beforeDir="false" afterPath="$PROJECT_DIR$/tables/password.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/tables/user.py" beforeDir="false" afterPath="$PROJECT_DIR$/tables/user.py" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/src/tables/usermember.py" beforeDir="false" afterPath="$PROJECT_DIR$/tables/usermember.py" afterDir="false" />
     </list>
     <option name="SHOW_DIALOG" value="false" />
     <option name="HIGHLIGHT_CONFLICTS" value="true" />
@@ -40,16 +47,16 @@
     <option name="hideEmptyMiddlePackages" value="true" />
     <option name="showLibraryContents" value="true" />
   </component>
-  <component name="PropertiesComponent"><![CDATA[{
-  "keyToString": {
-    "Python.dev-team-funcs.executor": "Run",
-    "Python.main.executor": "Run",
-    "RunOnceActivity.OpenProjectViewOnStart": "true",
-    "RunOnceActivity.ShowReadmeOnStart": "true",
-    "last_opened_file_path": "/home/bilge/development/hindsite",
-    "settings.editor.selected.configurable": "com.jetbrains.python.configuration.PyIntegratedToolsModulesConfigurable"
+  <component name="PropertiesComponent">{
+  &quot;keyToString&quot;: {
+    &quot;Python.dev-team-funcs.executor&quot;: &quot;Run&quot;,
+    &quot;Python.main.executor&quot;: &quot;Run&quot;,
+    &quot;RunOnceActivity.OpenProjectViewOnStart&quot;: &quot;true&quot;,
+    &quot;RunOnceActivity.ShowReadmeOnStart&quot;: &quot;true&quot;,
+    &quot;last_opened_file_path&quot;: &quot;/home/bilge/development/hindsite&quot;,
+    &quot;settings.editor.selected.configurable&quot;: &quot;com.jetbrains.python.configuration.PyIntegratedToolsModulesConfigurable&quot;
   }
-}]]></component>
+}</component>
   <component name="PyConsoleOptionsProvider">
     <option name="myPythonConsoleState">
       <console-settings module-name="hindsite" is-module-sdk="true">
@@ -60,9 +67,10 @@
   </component>
   <component name="RecentsManager">
     <key name="MoveFile.RECENT_KEYS">
+      <recent name="$PROJECT_DIR$/app" />
+      <recent name="$PROJECT_DIR$" />
       <recent name="$PROJECT_DIR$/src" />
       <recent name="$PROJECT_DIR$/src/app" />
-      <recent name="$PROJECT_DIR$" />
     </key>
   </component>
   <component name="RunManager" selected="Python.main">
@@ -102,7 +110,7 @@
       <option name="IS_MODULE_SDK" value="true" />
       <option name="ADD_CONTENT_ROOTS" value="true" />
       <option name="ADD_SOURCE_ROOTS" value="true" />
-      <option name="SCRIPT_NAME" value="$PROJECT_DIR$/src/main.py" />
+      <option name="SCRIPT_NAME" value="$PROJECT_DIR$/app/main.py" />
       <option name="PARAMETERS" value="" />
       <option name="SHOW_COMMAND_LINE" value="false" />
       <option name="EMULATE_TERMINAL" value="false" />
@@ -147,7 +155,7 @@
     <breakpoint-manager>
       <breakpoints>
         <line-breakpoint enabled="true" suspend="THREAD" type="python-line">
-          <url>file://$PROJECT_DIR$/src/main.py</url>
+          <url>file://$PROJECT_DIR$/app/main.py</url>
           <line>47</line>
           <option name="timeStamp" value="1" />
         </line-breakpoint>
